+++ default.nix
@@ -6,8 +6,7 @@
 let auto = fetchGit {
   url = "https://github.com/coq-community/coq-nix-toolbox.git";
   ref = "master";
-# putting a ref here is strongly advised
   rev = import .nix/coq-nix-toolbox.nix;
 };
 in
-(import auto ({inherit src;} // args)).nix-auto
+import auto ({inherit src;} // args)