+++ Changelog.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## UNRELEASED
+
+Compatible with
+- Coq 8.13 with Coq-Elpi 1.10.2.
+
+### General
+
+- Check for instances which break Forgetful Inheritance, attribute
+  `#[non_forgetful_inheritance]` to disable the check.
+- Do not impose useless universe constraints on `option` and `prod` by using
+  custom inductive types.
+- Attributes `#[primitive]` and `#[primitive_class]` for
+  `HB.structure/mixin/factory` to generate primitive records.
+- `Strategy Opaque` for named mixins
+- Factory instances are canonically (key `Factory.sort`) instances of all
+  the structures they can fulfill. This can be used inside proofs to provide
+  canonical instances on a type.
+- Attribute `#[doc="text"]` supported by all commands and used by `HB.about`
+### Commands
+
+- `HB.locate` and `HB.about`
+
 ## [1.1.0] - 2021-03-30
 
 Compatible with